adding delete user functionality

+++ frontend/src/context/storeContext.jsx
@@ -16,7 +16,7 @@ export const ContextProvider = ({ children }) => {
     const [editFormShow, setEditFormShow] = useState(false);
     const [user, setUser] = useState(null);
     const [users, setUsers] = useState([]);
-    const URL = "http://192.168.8.221:4002"
+    const URL = "http://localhost:4002"
     const token = localStorage.getItem("token");
 
     // decode function
@@ -124,12 +124,36 @@ export const ContextProvider = ({ children }) => {
     // delete jobs 
     const deleteJobs = async (id) => {
         try {
+            if(!window.confirm('Are you want to delete this job?')) return;
             const response = await axios.delete(`${URL}/api/jobs/delete-job/${id}`)
             const data = response.data;
             alert(data.message);
             fetchJobs();
         } catch (error) {
             console.error("faild to delete", error.massage)
+        }
+    }
+
+
+
+    // delete user account
+    const handleDeleteAccount = async () => {
+        if (!window.confirm('Are you want delete your account?')) return
+
+        try {
+            const config = {
+                headers: {
+                    Authorization: `Bearer ${localStorage.getItem('token')}`,
+                },
+            };
+
+            const res = await axios.delete(`${URL}/api/user/delete-account`, config);
+
+            alert(res.data.message);
+            localStorage.removeItem('token');
+            window.location.href = '/';
+        } catch (error) {
+            alert(error.response?.data?.message || 'Error deleting account')
         }
     }
     const contextValue = {
@@ -153,8 +177,9 @@ export const ContextProvider = ({ children }) => {
         setUsers,
         deleteUser,
         deleteJobs,
-        editFormShow, 
+        editFormShow,
         setEditFormShow,
+        handleDeleteAccount,
     }
     return (
         <StoreContext.Provider value={contextValue}>
